Standardize the layout of a task
