+++ src/covid_model_seiir_pipeline/forecasting/model.py
@@ -109,3 +109,39 @@ class ODERunner:
             columns=["S", "E", "I1", "I2", "R", "t", "beta", "theta"]
         )
         return result
+
+
+def get_ode_init_cond(location_id, beta_ode_fit, current_date):
+    """
+    Get the initial condition for the ODE.
+
+    Args:
+        location_id (init):
+            Location ids.
+        beta_ode_fit (str | pd.DataFrame):
+            The result for the beta_ode_fit, either file or path to file.
+        current_date (str | np.datetime64):
+            Current date for each location we try to predict off. Either file
+            or path to file.
+
+
+    Returns:
+         pd.DataFrame: Initial conditions by location.
+    """
+    # process input
+    assert (static_vars.COL_GROUP in beta_ode_fit)
+    assert (static_vars.COL_DATE in beta_ode_fit)
+    beta_ode_fit = beta_ode_fit[beta_ode_fit[static_vars.COL_GROUP] == location_id].copy()
+
+    if isinstance(current_date, str):
+        current_date = np.datetime64(current_date)
+    else:
+        assert isinstance(current_date, np.datetime64)
+
+    dt = np.abs((pd.to_datetime(beta_ode_fit[static_vars.COL_DATE]) - current_date).dt.days)
+    beta_ode_fit = beta_ode_fit.iloc[np.argmin(dt)]
+
+    col_components = static_vars.SEIIR_COMPARTMENTS
+    assert all([c in beta_ode_fit for c in col_components])
+
+    return beta_ode_fit[col_components].values.ravel()
